agents/python/bodysafety/requirements.txt
google-adk
pyyaml>=6.0
fastapi
uvicorn
python-multipart
google-genai

